# net/http: reduce memory usage when hijacking