@das7pad 2024-12-10
> It is possible to preserve the extended behavior of letting CloseNotify
> signal after the handler has exited by keeping around the relevant state
> for closeNotify in a struct that is used in both "connReader" and the
> new "connCloseNotify". It does not seem worth the extra code complexity
> for the deprecated CloseNotifier interface.

2bc45c12bea55251dd4c26c4eee4ed8d4f31241c is a WIP implementation for this -- in case this behavior is deemed useful to keep/worth the complexity.